Use gnumabs.
2002-10-14  Morten Welinder  <terra@diku.dk>

	* src/regression.c (non_linear_regression): Use gnumabs.

	* src/mathfunc.c (random_logistic): Use loggnum.

	* src/regression.c (parameter_errors): Use sqrtgnum.

	* src/format.c (fmt_general_int): Handle -2^31 correctly.

	* src/datetime.c (datetime_serial_raw_to_serial): Use floorgnum,
	not floor.
	(datetime_serial_raw_to_seconds): Ditto.

2002-10-14  Morten Welinder  <terra@diku.dk>

	* analysis-tools.c (analysis_tool_ttest_neqvar_engine_run): Use
	gnumabs, not fabs.

2002-10-14  Morten Welinder  <terra@diku.dk>

	* options.c: Use powgnum, not pow.  Use gnumabs, not fabs.

2002-10-14  Morten Welinder  <terra@diku.dk>

	* functions.c (gnumeric_even): Use fmodgnum.
	(gnumeric_odd): Ditto.
	(gnumeric_mround): Ditto.

+++ src/format.c
@@ -1499,7 +1499,7 @@ fmt_general_int (int val, int col_width)
 		if (val < 0) {
 			/* leave space for minus sign */
 			col_width--;
-			log_val = ceil (log10 (-val));
+			log_val = ceil (log10 ((unsigned int)-val));
 		} else
 			log_val = (val > 0) ? ceil (log10 (val)) : 0;
